# swagger_api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Blockchain Insurance Api

# basePath prefixes all resource paths 
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /eqcontract/confirmInsurance:
    # binds a127 app logic to a route
    x-swagger-router-controller: insurance_controller
    post:
      tags:
        - customer
      description: " "
      # used as the method name of the controller
      operationId: confirm
      parameters:
        - name: confirmRequest
          in: body
          description: ""
          required: true
          schema: 
            $ref: "#/definitions/confirmRequest"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /eqcontract/triggerInsurance:
    # binds a127 app logic to a route
    x-swagger-router-controller: insurance_controller
    post:
      tags:
        - customer
      description: " "
      # used as the method name of the controller
      operationId: trigger
      parameters:
        - name: triggerRequest
          in: body
          description: ""
          required: true
          schema: 
            $ref: "#/definitions/triggerRequest"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/triggerRespone"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  confirmRequest:
   properties:
      yourAddress:
        type: string
      eqcontract:
        type: string
      ether:
        type: number
  triggerRequest:
    properties:
      yourAddress:
        type: string
      eqcontract:
        type: string
  triggerRespone:
    properties:
      ether:
        type: number
...